Port of The Synth to the Arduino micro/leonardo / Atmel Atmega 32u4
This a modification/rewrite of the original library, to suits my needs.

What changed:
The wave, enveloppe and pitch tables have been changed.
The code is organised int three (2+1) files, as a normal arduino
library.
Number of channels is now 8 instead of 4.
Modulation engine has been removed.
Notes are generated from a table that use MIDI pitchs.
Functions have been added that enables the user to know the last
channel that was played, or the one which the play is more advanced.